move rcon function to utils
