+++ src/bot/commands/save.go
@@ -3,10 +3,11 @@ package commands
 import (
 	"github.com/juunini/palworld-discord-bot/src/console_decoration"
 	"github.com/juunini/palworld-discord-bot/src/i18n"
+	"github.com/juunini/palworld-discord-bot/src/utils"
 )
 
 func save() string {
-	client, err := rconClient()
+	client, err := utils.RconClient()
 	if err != nil {
 		return i18n.FailedToConnectRconServer
 	}
